--- src/FamilyHubs.ServiceDirectory.Admin.Web/Areas/SubjectAccessRequest/Pages/Shared/_laConnectionStatus.cshtml
@@ -1,18 +1,0 @@
-﻿@using FamilyHubs.ServiceDirectory.Admin.Core.ApiClient
-@model FamilyHubs.ReferralService.Shared.Dto.ReferralDto
-
-@switch ((ReferralStatus)Model.Status.Id)
-{
-    case ReferralStatus.Opened:
-    case ReferralStatus.New:
-        <strong class="govuk-tag govuk-tag--purple">Sent</strong>
-        break;
-    case ReferralStatus.Accepted:
-        <strong class="govuk-tag govuk-tag--green">Accepted</strong>
-        break;
-    case ReferralStatus.Declined:
-        <strong class="govuk-tag govuk-tag--red">Declined</strong>
-        break;
-    default:
-        throw new ArgumentOutOfRangeException($"Unexpected status id {Model.Status.Id}");
-}
