--- src/FamilyHubs.ServiceDirectory.Admin.Web/Areas/SubjectAccessRequest/_Layout.cshtml
@@ -1,17 +1,0 @@
-﻿@{
-    Layout = "~/pages/Shared/_Layout.cshtml";
-}
-
-@section Head {
-    @await RenderSectionAsync("Head", false)
-}
-
-@section Scripts{
-    @await RenderSectionAsync("Scripts", false)
-}
-
-@section Back {
-    @await RenderSectionAsync("Back", required: false)
-}
-
-@RenderBody()
