--- src/FamilyHubs.ServiceDirectory.Admin.Web/Areas/SubjectAccessRequest/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@namespace FamilyHubs.ServiceDirectory.Admin.Web.Areas.SubjectAccessRequest
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-
