--- src/FamilyHubs.ServiceDirectory.Admin.Web/Areas/SubjectAccessRequest/_ViewStart.cshtml
@@ -1,8 +1,0 @@
-﻿@using FamilyHubs.SharedKernel.Razor.Layout
-@inject FamilyHubs.SharedKernel.Razor.FamilyHubsUi.FamilyHubsLayoutModel FamilyHubsLayoutModel
-@model Microsoft.AspNetCore.Mvc.RazorPages.PageModel
-
-@{
-    Layout = "./_Layout.cshtml";
-    ViewStart.InitialiseFamilyHubs(FamilyHubsLayoutModel, Model, ViewData);
-}
