+++ src/FamilyHubs.ServiceDirectory.Admin.Web/Pages/La/PersonsDetails.cshtml
@@ -9,7 +9,7 @@
 }
 
 @section Back {
-    <a asp-page="\Welcome" class="govuk-back-link">Back</a>
+    <a href="/Welcome" class="govuk-back-link">Back</a>
 }
 
 <div class="govuk-grid-row">
