+++ src/FamilyHubs.ServiceDirectory.Admin.Web/Pages/La/RequestDetails.cshtml
@@ -15,7 +15,7 @@
 }
 
 @section Back {
-    <a href="/SubjectAccessRequest/SubjectAccessResultDetails"  class="govuk-back-link">Back</a>
+    <a href="/La/SubjectAccessResultDetails"  class="govuk-back-link">Back</a>
 }
 
 <div class="govuk-grid-row">
