+++ src/FamilyHubs.ServiceDirectory.Admin.Web/Pages/La/SubjectAccessResultDetails.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model FamilyHubs.ServiceDirectory.Admin.Web.Areas.SubjectAccessRequest.Pages.SubjectAccessResultDetailsModel
+@model SubjectAccessResultDetailsModel
 @{
     ViewData["Title"] = "Subject Access Result Details?";
 }
@@ -9,7 +9,7 @@
 }
 
 @section Back {
-    <a asp-page="PersonsDetails" asp-area="SubjectAccessRequest" class="govuk-back-link">Back</a>
+    <a href="/La/PersonsDetails" class="govuk-back-link">Back</a>
 }
 
 <div class="govuk-grid-row">
