+++ src/FamilyHubs.ServiceDirectory.Admin.Web/Pages/Welcome.cshtml
@@ -140,7 +140,7 @@
 
             <div class="card">
                 <h3 class="govuk-heading-s">
-                    <a asp-area="SubjectAccessRequest" asp-page="/PersonsDetails">Subject Access Request</a>
+                    <a href="/La/PersonsDetails" >Subject Access Request</a>
                 </h3>
                 <p class="govuk-body-s">Find information stored about a person.</p>
             </div>
